requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
moviepy>=1.0.3
